Rename 'input-params' to 'configuration'
