+++ sources/src/dependency-submission/post.ts
@@ -1,7 +1,7 @@
 import * as core from '@actions/core'
 import * as setupGradle from '../setup-gradle'
 
-import {CacheConfig, SummaryConfig} from '../input-params'
+import {CacheConfig, SummaryConfig} from '../configuration'
 import {PostActionJobFailure} from '../errors'
 
 // Catch and log any unhandled exceptions.  These exceptions can leak out of the uploadChunk method in
